Separate repositories table
