+++ client/galaxy/scripts/components/Toolshed/Repositories.vue
@@ -1,0 +1,112 @@
+<template>
+    <div>
+        <b-table striped :items="repositories" :fields="fields">
+            <template slot="name" slot-scope="row">
+                <b-link href="#" class="font-weight-bold" @click="row.toggleDetails">
+                    {{ row.item.name }}
+                </b-link>
+            </template>
+            <template slot="row-details" slot-scope="row">
+                <repositoryoptions :repo="row.item" :toolSections="toolSections"/>
+            </template>
+        </b-table>
+        <div v-if="noResultsFound">
+            No matching repositories found.
+        </div>
+        <div v-if="pageLoading">
+            <span class="fa fa-spinner fa-spin mb-4" /> <span>Loading repositories...</span>
+        </div>
+    </div>
+</template>
+<script>
+import Vue from "vue";
+import BootstrapVue from "bootstrap-vue";
+import { getAppRoot } from "onload/loadConfig";
+import { getGalaxyInstance } from "app";
+import RepositoryOptions from "./RepositoryOptions.vue";
+import { Services } from "./services.js";
+const READY = 0;
+const LOADING = 1;
+const COMPLETE = 2;
+export default {
+    components: {
+        repositoryoptions: RepositoryOptions
+    },
+    props: ["query", "scrolled"],
+    data() {
+        return {
+            toolshedUrl: "https://toolshed.g2.bx.psu.edu/",
+            toolSections: [],
+            repositories: [],
+            fields: [
+                { key: "name" },
+                { key: "description" },
+                { key: "last_updated", label: "Updated" },
+                { key: "repo_owner_username", label: "Owner" },
+                { key: "times_downloaded", label: "Downloaded" }
+            ],
+            page: 1,
+            pageSize: 10,
+            pageState: READY,
+            error: null
+        };
+    },
+    watch: {
+        query() {
+            this.load();
+        },
+        scrolled() {
+            if (this.scrolled && this.pageState === READY) {
+                this.load(this.page + 1);
+            }
+        }
+    },
+    computed: {
+        pageLoading() {
+            return this.pageState === LOADING;
+        },
+        noResultsFound() {
+            return this.repositories.length === 0 && !this.pageLoading;
+        }
+    },
+    created() {
+        this.services = new Services();
+        this.setToolSections();
+        this.load();
+    },
+    methods: {
+        setToolSections() {
+            const galaxy = getGalaxyInstance();
+            const sections = galaxy.config.toolbox_in_panel;
+            this.toolSections = sections.filter(x => x.model_class == "ToolSection").map(x => x.name);
+        },
+        load(page = 1) {
+            this.page = page;
+            this.pageState = LOADING;
+            this.services
+                .getRepositories({
+                    tool_shed_url: this.toolshedUrl,
+                    q: this.query,
+                    page: this.page,
+                    page_size: this.pageSize
+                })
+                .then(incoming => {
+                    if (this.page === 1) {
+                    this.repositories = incoming;
+                    } else {
+                        this.repositories = this.repositories.concat(incoming);
+                    }
+                    if (incoming.length < this.pageSize) {
+                        this.pageState = COMPLETE;
+                    } else {
+                        this.pageState = READY;
+                    }
+                    this.error = null;
+                })
+                .catch(errorMessage => {
+                    this.error = errorMessage;
+                });
+        }
+    }
+};
+</script>
